Set default clonedir for Git upstreams to "."
For a Git upstream, we should set the default clone dir to also be
".", the same default directory for decompressing the contents of a
plain upstream to. This way we ensure consistent location/behavior
between plain and Git upstreams when only one upstream is present. If
there are more than one, packagers have to specify the `clonedir`
just like how they specify `unpackdir` for plain upstreams, otherwise
files from different upstreams may clash with each other.

Signed-off-by: Gavin Zhao <git@gzgz.dev>